Extract common test functionality
